Create clusters of fixed size

Create clusters of fixed size using the nearest neighbor approach.

Usage

fixed_size_clustering_nearest_neighbor(d, size, method = c("max", "min"))

fixed_size_clustering_kmeans(x, size, method = c("range", "min", "max"))

fixed_size_clustering_hclust(d, size, method = "ward.D2")

Arguments

d

dist, dissimilarity structure as produced by dist.

size

integer, fixed target size of the clusters.

method

character, start with the closest (min)/farthest (max) point in average.

x

numeric/matrix, points to be clustered.

Value

integer, vector of the same length as attr(d, "Size") with cluster identification numbers.

Details

fixed_size_clustering_kmeans: This algorithm runs [stats::kmeans()] once to determine the center position and reassigns the points to the nearest center with respect to the cluster size. The kmeans could run multiple times to update every center after a cluster was filled but we used the current approach for simplicity.

methods: min (max), start with the closest (farthest) points. range start with the points where the difference in farthest minus nearest distance is largest.

fixed_size_clustering_hclust: This algorithm runs [stats::hclust()] once to determine the clustering and cuts the tree at different heigths to ensure the fixed cluster sizes.

methods: ward.D2 clustering method, see [stats::hclust()] for details.

References

Algorithm taken from Jean Monlong https://github.com/jmonlong/Hippocamplus/blob/master/content/post/2018-06-09-ClusterEqualSize.Rmd

kmeans Algorithm inspired from from Jean Monlong https://github.com/jmonlong/Hippocamplus/blob/master/content/post/2018-06-09-ClusterEqualSize.Rmd and https://elki-project.github.io/tutorial/same-size_k_means

Examples

x <- c(1:3, 1e4, 11:13, 1.1e4, 111:113)
d <- dist(x)
fixed_size_clustering_nearest_neighbor(d, size = 3)
#>  [1] 3 3 3 1 4 4 4 1 2 2 2
fixed_size_clustering_kmeans(d, size = 3)
#>  [1] 2 2 2 4 1 1 1 4 3 3 3
fixed_size_clustering_hclust(d, size = 3)
#>  [1] 1 1 1 4 2 2 2 4 3 3 3